refactor: separate selectors into a module
