+++ src/store/api-actions.js
@@ -1,23 +1,30 @@
 import {loadFilms, loadPromo, requireAuthorization} from "./action";
 import {AuthorizationStatus, APIRoute} from "../consts";
 
-export const fetchFilms = () => (dispatch, _getState, api) => (
+const fetchFilms = () => (dispatch, _getState, api) => (
   api.get(APIRoute.FILMS)
     .then(({data}) => dispatch(loadFilms(data)))
 );
 
-export const fetchPromo = () => (dispatch, _getState, api) => (
+const fetchPromo = () => (dispatch, _getState, api) => (
   api.get(APIRoute.PROMO)
     .then(({data}) => dispatch(loadPromo(data)))
 );
 
-export const checkAuth = () => (dispatch, _getState, api) => (
+const checkAuth = () => (dispatch, _getState, api) => (
   api.get(APIRoute.LOGIN)
     .then(() => dispatch(requireAuthorization(AuthorizationStatus.AUTH)))
     .catch(() => {})
 );
 
-export const login = ({login: email, password}) => (dispatch, _getState, api) => (
+const login = ({login: email, password}) => (dispatch, _getState, api) => (
   api.post(APIRoute.LOGIN, {email, password})
     .then(() => dispatch(requireAuthorization(AuthorizationStatus.AUTH)))
 );
+
+export {
+  fetchFilms,
+  fetchPromo,
+  login,
+  checkAuth,
+};
