+++ src/store/reducers/data/selectors.js
@@ -1,0 +1,3 @@
+const selectData = (state) => state.DATA;
+
+export {selectData};
