+++ src/store/reducers/filter/selectors.js
@@ -1,0 +1,3 @@
+const selectFilter = (state) => state.FILTER;
+
+export {selectFilter};
