+++ src/store/reducers/user/selectors.js
@@ -1,0 +1,3 @@
+const selectUser = (state) => state.USER;
+
+export {selectUser};
